104-fibonacci

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -1,4 +1,5 @@
 #include <stdio.h>
+#include <stdint.h>
 
 /**
  * main - print the first 98 Fibonacci numbers
@@ -10,22 +11,22 @@
 int main(void)
 {
 	int i;
-	long prev = 1;
-	long cur = 2;
-	long result = 0;
+	__int128 prev = 1;
+	__int128 cur = 2;
+	__int128 result = 0;
 
 	for (i = 1; i <= 98; i++)
 	{
 		if (i == 1)
-			printf("%lu, ", prev);
+			printf("%d, ", prev);
 		else if (i == 2)
-			printf("%lu", cur);
+			printf("%d", cur);
 		else
 		{
 			result = (prev + cur);
 			prev = cur;
 			cur = result;
-			printf(", %lu", result);
+			printf(", %d", result);
 		}
 	}
